% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fundamental_diagram.R
\name{calculate_axes}
\alias{calculate_axes}
\title{Calculate necessary data to trace the fundamental diagram}
\usage{
calculate_axes(enriched_data, sensor_version = 1, direction_choice = NULL)
}
\arguments{
\item{enriched_data}{enriched data.frame containing all the data for all your sensors}

\item{sensor_version}{Integer (1,2). Selected the version of the sensor (data from Telraam). Default to 1.}

\item{direction_choice}{Character vector. Selected direction ('rgt','lft'). Both by default (NULL)}
}
\value{
enriched_data
}
\description{
Calculate necessary data to trace the fundamental diagram
}
\examples{
calculate_axes(traffic)
calculate_axes(traffic,
  sensor_version=2,
  direction_choice='rgt')
}
