% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{check_options_graph}
\alias{check_options_graph}
\title{Check if options are available in the options list, replace by a default otherwise.}
\usage{
check_options_graph(options_selected, options_available, default)
}
\arguments{
\item{options_selected}{List of characters. Selected options.}

\item{options_available}{List of characters. Valid options.}

\item{default}{List of characters. Default options.}
}
\value{
Options consistent with the possibilities
}
\description{
Check if options are available in the options list, replace by a default otherwise.
}
\keyword{internal}
