% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Create config file if needed.}
\usage{
create_config(overwrite = FALSE)
}
\arguments{
\item{overwrite}{If the file exist, should it be overwriten?}
}
\value{
Boolean, TRUE if the file is created, FALSE overwise (config already exists for example).
}
\description{
If the user needs to specify the IDs of his own sensors,
this function create a local configuration template file
in the \verb{\\inst} directory, to edit with specific information.
}
\examples{
create_config() #doesn't work: a config file already exists.

}
