% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{custom_colors_palette}
\alias{custom_colors_palette}
\title{Colors palettes for each option}
\usage{
custom_colors_palette(segments)
}
\arguments{
\item{segments}{List of characters. Segments name of the dataframe.}
}
\value{
list of color palettes (named vector) for each option
}
\description{
Colors palettes for each option
}
\keyword{internal}
