% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_decompose_2_fct.R
\name{decompose_2_fct}
\alias{decompose_2_fct}
\title{Decompose data from two sensors}
\usage{
decompose_2_fct(
  data,
  sensor1,
  sensor2,
  hour_x,
  direction1,
  direction2,
  mobility,
  norm,
  vacations = NULL,
  public_holidays = NULL
)
}
\arguments{
\item{data}{Data frame containing the data}

\item{sensor1}{The segment ID for sensor 1}

\item{sensor2}{The segment ID for sensor 2}

\item{hour_x}{The hour to filter on}

\item{direction1}{The direction for sensor 1}

\item{direction2}{The direction for sensor 2}

\item{mobility}{A character vector indicating the mobility types to include in the filter}

\item{norm}{A character indicating whether to normalize the results}

\item{vacations}{vacations period for filtering function}

\item{public_holidays}{public holidays period for filtering function}
}
\value{
A list containing the decomposed data, seasonal data, peaks, and correlation
}
\description{
Decompose data from two sensors
}
