% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrich.R
\name{enrich_dates}
\alias{enrich_dates}
\title{Enrich traffic data with date informations.}
\usage{
enrich_dates(data)
}
\arguments{
\item{data}{Data frame containing a \code{date} character column containing date + hour + timezone.}
}
\value{
Same dataframe with 3 additionnal columns : day, weekday and hour.
}
\description{
Enrich traffic data with date informations.
}
\examples{
df <- data.frame('date' = c("2022-02-18 08:00:00 CET","2022-02-18 09:00:00 CET"),
  stringsAsFactors = FALSE)
enrich_dates(df)

}
\keyword{internal}
