% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_enrich_data.R
\name{enrich_name}
\alias{enrich_name}
\title{Enrich traffic data with segment name
Segment fullname is also added : it's the combination of segment's id and name.}
\usage{
enrich_name(data)
}
\arguments{
\item{data}{Data frame containing a segment_id column}
}
\value{
Same dataframe with two additionnal columns : segment name and full name
}
\description{
Enrich traffic data with segment name
Segment fullname is also added : it's the combination of segment's id and name.
}
\keyword{internal}
