% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_enrich_data.R
\name{enrich_special_days}
\alias{enrich_special_days}
\title{Enrich traffic data with french vacation and public holidays}
\usage{
enrich_special_days(data, vacations = NULL, public_holidays = NULL)
}
\arguments{
\item{data}{Data frame containing a uptime column}

\item{vacations}{data frame containing the vacation dates}

\item{public_holidays}{data frame containing the public holidays dates}
}
\value{
Same dataframe with two additionnal boolean columns (TRUE if the day corresponds to vacation/public holiday)
}
\description{
Enrich traffic data with french vacation and public holidays
}
\keyword{internal}
