% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_enrich.R
\name{enrich_traffic}
\alias{enrich_traffic}
\title{Enrich traffic data with date features, names and uptime filters.
This function add day, weekday, hour, segment_name and full_name and uptime quality boolean.}
\usage{
enrich_traffic(data)
}
\arguments{
\item{data}{Raw data frame from the Telraam API, imported through the package.}
}
\value{
Same dataframe with additionnal informations.
}
\description{
Enrich traffic data with date features, names and uptime filters.
This function add day, weekday, hour, segment_name and full_name and uptime quality boolean.
}
\keyword{internal}
