% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_enrich_data.R
\name{enrich_uptime}
\alias{enrich_uptime}
\title{Enrich traffic data with uptime quality indication
If the uptime is lower than 0.5, uptime_quality will be FALSE, else TRUE}
\usage{
enrich_uptime(data)
}
\arguments{
\item{data}{Data frame containing a uptime column}
}
\value{
Same dataframe with an additionnal column indicating if the uptime is greater or lower than 0.5.
}
\description{
Enrich traffic data with uptime quality indication
If the uptime is lower than 0.5, uptime_quality will be FALSE, else TRUE
}
\keyword{internal}
