% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fundamental_diagram.R
\name{filter_demand_user}
\alias{filter_demand_user}
\title{Filter the dataframe from user's demand}
\usage{
filter_demand_user(
  enriched_data,
  segments,
  date_range,
  weekday_choice,
  hour_choice,
  vacation_choice,
  holiday_choice
)
}
\arguments{
\item{enriched_data}{enriched data.frame containing all the data for all your sensors}

\item{segments}{Character vector. Selected road segment, all if NULL (default).}

\item{date_range}{Date vector. example: c('2021-01-01','2022-01-01'). Full period if NULL (default).}

\item{weekday_choice}{weekday Character vector. Weekday choosen. Default to the all week.}

\item{hour_choice}{Integer vector. Hours choosen, default to the all day.}

\item{vacation_choice}{Character vector. Selected vacation. Full period by default (NULL).}

\item{holiday_choise}{Boolean. Selected holiday.  Full period by default (NULL).}
}
\value{
enriched_data
}
\description{
Filter the dataframe from user's demand
}
\examples{
plot_fundamental_diagramm(traffic)
plot_fundamental_diagramm(traffic,
  date_range = c('2022-07-01','2022-09-01'),
  weekday_choice= c('monday','friday','sunday'),
  hour_choice= c(1,5,10,14,21),
  vacation_choice=NULL,
  holiday_choice=TRUE,
  segments = 'RteVitre-06')
}
