% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{filtering_agg}
\alias{filtering_agg}
\title{Filter by selected criteria and aggregating traffics.}
\usage{
filtering_agg(
  data,
  date_range = NULL,
  segments = NULL,
  direction = NULL,
  modes = NULL
)
}
\arguments{
\item{data}{Traffic Data Frame}

\item{date_range}{Date vector, c("aaaa-mm-jj","aaaa-mm-jj")}

\item{direction}{Vector of character. Direction of the street (lft, right, both).}

\item{modes}{Vector of character. Type(s) of mobility: c("car","heavy","pedestrian","bike")#'}

\item{segment}{Vector of character. Ids of desired segments.}
}
\value{
the filtered data with a new column "traffic" with aggregated data for specified direction/modes
}
\description{
Not all criteria need to be filled in. Unfilled criteria are set by default so that no filtering is performed.
}
\keyword{internal}
