% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{get_config_path}
\alias{get_config_path}
\title{Get the path of configuration file.}
\usage{
get_config_path()
}
\value{
Character: path of the configuration file, needed for a lot of functions.
}
\description{
The configuration file could be in the \code{inst} directory, or in a temporary
directory, according to the \code{create_directory} option of the \code{create_config()}
function. If the configuration file doesn't exist, this function create a file
in a temporary directory and send a message to the user.
}
\examples{
create_config(create_directory=FALSE)
get_config_path()
}
\keyword{internal}
