% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graph.R
\name{get_custom_palette}
\alias{get_custom_palette}
\title{Colors palettes for each option}
\usage{
get_custom_palette(segments)
}
\arguments{
\item{segments}{List of characters. Segments name of the dataframe.}
}
\value{
list of color palettes (named vector) for each option
}
\description{
Colors palettes for each option
}
\keyword{internal}
