% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_global.R
\name{get_segment_name}
\alias{get_segment_name}
\title{Get the name of a segment giving its id}
\usage{
get_segment_name(segment_id)
}
\arguments{
\item{segment_id}{ID of segment, should be present in inst/config.yml}
}
\value{
Name of the segment, as specified in the configuration file, NULL otherwise.
}
\description{
Get the name of a segment giving its id
}
\examples{
\dontrun{ #run if you want to create a inst/ directory containing config file
  create_config()
  get_segment_name(9000000000)
  }
}
\keyword{internal}
