% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_api_state.R
\name{get_telraam_token}
\alias{get_telraam_token}
\title{Get the current authentication Token for the telraam API}
\usage{
get_telraam_token()
}
\value{
Token currently used
}
\description{
Get the current authentication Token for the telraam API
}
