% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_gg_na_heatmap.R
\name{gg_na_heatmap}
\alias{gg_na_heatmap}
\title{gg_na_heatmap
This function generates a heatmap representation of the proportion of missing values (NA) for different sensors over monthly periods during a selected time frame.}
\usage{
gg_na_heatmap(
  data,
  start = NULL,
  end = NULL,
  hours = "ALL",
  list_weekday = "ALL",
  pub_holidays = "YES",
  holidays = "YES"
)
}
\arguments{
\item{data}{Data frame. A data frame containing road traffic data.}

\item{start}{Character. The start date in the format "yyyy-mm-dd".}

\item{end}{Character. The end date in the format "yyyy-mm-dd".}

\item{hours}{Vector. A vector containing hours on which you want to see the representation, e.g., c("05h", "18h").}

\item{list_weekday}{Vector. A vector containing weekdays in French, e.g., c("Monday", "Sunday").}

\item{pub_holidays}{Character. "YES" if you want to see results with public holidays included; "NO" if you want to exclude them from the representation; "ONLY" if you only want to see them in the result.}

\item{holidays}{Character. "YES" if you want to see results including holidays; "NO" if you want to exclude them from the representation; "ONLY" if you only want to see them in the result.}
}
\value{
A ggplot graph representing the heatmap of NA proportions for different sensors over monthly periods.
}
\description{
gg_na_heatmap
This function generates a heatmap representation of the proportion of missing values (NA) for different sensors over monthly periods during a selected time frame.
}
