% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_plot_evolution.R
\name{gg_traffic_evolution}
\alias{gg_traffic_evolution}
\title{Evolution of traffic (global, per mode ou per direction), smoothed traffic
during a period.}
\usage{
gg_traffic_evolution(
  enriched_data,
  date_range = NULL,
  segments = NULL,
  modes = c("heavy", "car"),
  direction = "both",
  smoothed = TRUE,
  agg_day = TRUE
)
}
\arguments{
\item{enriched_data}{enriched data.frame containing all the data for all your sensors}

\item{date_range}{Date vector. example: c('2021-01-01','2022-01-01'). Full period if NULL (default).}

\item{segments}{Character vector. Selected road segment by its name, all if NULL (default).}

\item{modes}{Character vector. Different modes of transportation aggregated (heavy, car, bike, pedestrian) . Default: heavy & car}

\item{direction}{Character. Direction of the traffic (lft, rgt, both). Default to both.}

\item{smoothed}{Boolean. Should the smoothed traffic be plotted ? Default: True}

\item{agg_day}{Boolean. Should the data be aggregated per day ? Default : True}
}
\value{
graph showing the evolution of traffic (for specified parameters) during the specified period.
}
\description{
Evolution of traffic (global, per mode ou per direction), smoothed traffic
during a period.
}
