% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{graph_subtitles}
\alias{graph_subtitles}
\title{Create subtitles for graphics.}
\usage{
graph_subtitles(
  segments = NULL,
  modes = NULL,
  directions = NULL,
  weekdays = NULL,
  hours = NULL
)
}
\arguments{
\item{segments}{Character vector. Selected road segment, all if NULL (default).}

\item{modes}{Character vector. Different modes of transportation aggregated (heavy, car, bike, pedestrian) . Default: heavy & car}

\item{directions}{Character vector. Direction of the traffic (lft, rgt, both). Default to both.}

\item{weekdays}{Character vector. Weekday choosen. Default to the all week.}
}
\value{
DataFrame with one row per hour/segment/transportation mode/direction
}
\description{
Create subtitles for graphics.
}
\keyword{internal}
