% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_plot_average.R
\name{melt_direction_mode}
\alias{melt_direction_mode}
\title{Melt traffic dataframe to create mode and transportation columns to help graphic representation}
\usage{
melt_direction_mode(data, modes, direction = TRUE, agg_mode = FALSE)
}
\arguments{
\item{data}{enriched data.frame containing all the data for all your sensors}

\item{modes}{Character vector. Different modes of transportation aggregated (heavy, car, bike, pedestrian) . Default: heavy & car}

\item{direction}{Boolean. Default to TRUE : direction is mandatory for all modes.}

\item{agg_mode}{Boolean. Should all modes be aggregated ? Default to FALSE.}
}
\value{
Molten DataFrame with one or two new columns : mode + transportation if required.
}
\description{
Melt traffic dataframe to create mode and transportation columns to help graphic representation
}
\keyword{internal}
