% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fundamental_diagram.R
\name{plot_fundamental_diagramm}
\alias{plot_fundamental_diagramm}
\title{Plot three fundamentals diagrams}
\usage{
plot_fundamental_diagramm(
  enriched_data,
  date_range = NULL,
  weekday_choice = NULL,
  hour_choice = NULL,
  vacation_choice = NULL,
  holiday_choice = NULL,
  segments = NULL,
  direction_choice = NULL
)
}
\arguments{
\item{enriched_data}{enriched data.frame containing all the data for all your sensors}

\item{date_range}{Date vector. example: c('2021-01-01','2022-01-01'). Full period if NULL (default).}

\item{weekday_choice}{weekday Character vector. Weekday choosen. Default to the all week.}

\item{hour_choice}{Integer vector. Hours choosen, default to the all day.}

\item{vacation_choice}{Character vector. Selected vacation. Full period by default (NULL).}

\item{segments}{Character vector. Selected road segment, all if NULL (default).}

\item{direction_choice}{Character vector. Selected direction ('rgt','lft'). Both by default (NULL)}

\item{holiday_choise}{Boolean. Selected holiday.  Full period by default (NULL).}
}
\value{
enriched_data
}
\description{
Plot three fundamentals diagrams
}
\examples{
plot_fundamental_diagramm(traffic)
plot_fundamental_diagramm(traffic,
  date_range = c('2022-07-01','2022-09-01'),
  weekday_choice = c('monday','friday','sunday'),
  hour_choice = c(1,5,10,14,21),
  holiday_choice = TRUE,
  segments = 'RteVitre-06',
  direction_choice = 'rgt')
}
