% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fundamental_diagram.R
\name{recons_v85}
\alias{recons_v85}
\title{Reconstitute the v85 by direction from speed_hist_car}
\usage{
recons_v85(df)
}
\arguments{
\item{df}{to fill}
}
\value{
to fill
}
\description{
Add two colums in the dataframe : v85_lft and v85_rgt
}
\examples{
#to fill
}
