% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{retrieve_missing_data}
\alias{retrieve_missing_data}
\title{Clean data from inactivity periods and missing hours}
\usage{
retrieve_missing_data(
  data,
  date_range = NULL,
  segment_id = NULL,
  threshold_uptime = 0.5,
  successive_day = 2,
  remove_data = TRUE,
  show_graph = TRUE
)
}
\arguments{
\item{data}{data.frame containing all the data for all your sensors}

\item{date_range}{Date vector. Example: c('2021-01-01','2022-01-01'). Full period if NULL (default).}

\item{segment_id}{Character vector. Selected road segment, all if NULL (default).}

\item{threshold_uptime}{Numeric. Uptime threshold chosen. Default is 0.5.}

\item{successive_day}{Integer. Number of days chosen to define an inactivity period. Default is 2.}

\item{remove_data}{Logical. If TRUE, completely removes inactivity periods. If FALSE, replaces data with NA.}

\item{show_graph}{Logical. If TRUE, displays a graph of inactivity periods. Default is TRUE.}
}
\value{
Cleaned data
}
\description{
This function cleans the data by removing nighttime hours and periods of inactivity.
It combines the functionalities of \code{retrieve_missing_hours} and \code{replace_inactivity_period}.
}
\examples{
traffic_cleaned <- retrieve_missing_data(traffic)
traffic_cleaned <- retrieve_missing_data(traffic,
                                         date_range = c('2022-07-01','2022-09-01'),
                                         segment_id = 9000001844,
                                         threshold_uptime = 0.3,
                                         successive_day = 1,
                                         remove_data = FALSE,
                                         show_graph = TRUE)
}
