% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{retrieve_missing_hours}
\alias{retrieve_missing_hours}
\title{Remove nighttime hours from traffic data}
\usage{
retrieve_missing_hours(data, threshold_uptime = 0.5)
}
\arguments{
\item{data}{data.frame containing all the data for all your sensors}

\item{threshold_uptime}{Numeric. Uptime threshold chosen. Default is 0.5.}
}
\value{
data without nighttime hours
}
\description{
This function removes nighttime hours when sensors cannot capture information.
It filters the data to retain only the hours with reliable uptime.
}
\examples{
traffic_without_nighttime <- retrieve_missing_hours(traffic)
traffic_without_nighttime <- retrieve_missing_hours(traffic, threshold_uptime = 0.3)
}
