% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_retrieve_sensor.R
\name{retrieve_sensor}
\alias{retrieve_sensor}
\title{Retrieves data associated with a sensor from the Telraam API}
\usage{
retrieve_sensor(segment_name, start_date, end_date, key = get_telraam_token())
}
\arguments{
\item{segment_name}{Character. Name of the segment, as specified in config.}

\item{start_date}{Date. Start date "aaaa-mm-jj", must be of the date type.}

\item{end_date}{Date. End date "aaaa-mm-jj", must be of the date type.}

\item{key}{the api key (set by the set_telraam_token function)}
}
\description{
Retrieves data associated with a sensor from the Telraam API.
The data is retrieved for a specified time period between \code{start_date} and \code{end_date} (inclusive).
}
