% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrieve.R
\name{seq_by_3_month}
\alias{seq_by_3_month}
\title{Generate sequence of intervals with three-month periods}
\usage{
seq_by_3_month(start_date, end_date)
}
\arguments{
\item{start_date}{Date. Start date in "yyyy-mm-dd" format.}

\item{end_date}{Date. End date in "yyyy-mm-dd" format.}
}
\value{
Dataframe with a \code{start} and an \code{end} columns, each row represents 3 months period.
}
\description{
This function is used internally in the \code{retrieve_sensor} function to generate a sequence of intervals with three-month periods. It takes a start date (\code{start_date}) and an end date (\code{end_date}), and returns a data frame with two columns representing the start and end dates of each interval.
}
\examples{
seq_by_3_month(as.Date('2023-11-01'),as.Date('2024-11-01'))

}
\keyword{internal}
