% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_state.R
\name{set_telraam_token}
\alias{set_telraam_token}
\title{Saves an authentication token for the 'Telraam' API.}
\usage{
set_telraam_token(token)
}
\arguments{
\item{token}{a \code{string} with the token}
}
\value{
Boolean: TRUE if the token is correctly set
}
\description{
If you want to get this token after this instruction, please use \code{get_telraam_token()}.
}
\examples{
my_token <- "MyTelraamToken"
set_telraam_token(my_token)
get_telraam_token()

}
