% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_simple_plot.R
\name{simple_plot}
\alias{simple_plot}
\title{Number of cars and Heavies as a function of time}
\usage{
simple_plot(
  data,
  sensor,
  date_range,
  line_color = "black",
  smooth_color = "#B1D62E"
)
}
\arguments{
\item{data}{data.frame containing all the data for all your sensors}

\item{sensor}{numeric, Id of the chosen sensor}

\item{date_range}{Date vector. example: c('2021-01-01','2022-01-01')}

\item{line_color}{character. color of main line. Default black}

\item{smooth_color}{character. color of smoothing line. Default to light green}
}
\value{
graph showing the evolution of the number of cars and heavies over the selected date range for the chosen sensor.
}
\description{
Number of cars and Heavies as a function of time
}
