% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_utils.R
\name{which_vacations}
\alias{which_vacations}
\title{Indicates if a date is in vacation period and if true, which vacation.
If the date is not in a vacation period, "No vacation" is returned.}
\usage{
which_vacations(date, vacation)
}
\arguments{
\item{date}{Date (character format)}

\item{vacation}{Dataframe of vacations, same format as set_globals_vars output.
TODO : this function could be more efficient.}
}
\value{
Vacation description if the day is between two dates, "No vacation" otherwise.
}
\description{
Indicates if a date is in vacation period and if true, which vacation.
If the date is not in a vacation period, "No vacation" is returned.
}
\keyword{internal}
