% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graph.R
\name{check_options_graph}
\alias{check_options_graph}
\title{Check if options are available in the options list, replace by a default otherwise.}
\usage{
check_options_graph(options_selected, options_available, default)
}
\arguments{
\item{options_selected}{List of characters. Selected options.}

\item{options_available}{List of characters. Valid options.}

\item{default}{List of characters. Default options.}
}
\value{
Options consistent with the possibilities
}
\description{
Check if options are available in the options list, replace by a default otherwise.
}
\examples{
check_options_graph(c('car','pedestrian'),
  c('car','pedestrian','bike','heavy'),c('car','heavy'))
check_options_graph(c('coucou','salut'),
  c('car','pedestrian','bike','heavy'),c('car','heavy'))
check_options_graph(NULL,
  c('car','pedestrian','bike','heavy'),c('car','heavy'))

}
\keyword{internal}
