% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_data.R
\name{create_and_train_model}
\alias{create_and_train_model}
\title{TODO}
\usage{
create_and_train_model(data, target, base_vars, threshold_uptime)
}
\arguments{
\item{data}{TBD}

\item{target}{TBD}

\item{base_vars}{TBD}

\item{threshold_uptime}{TBD}
}
\value{
TBD
}
\description{
TODO
}
\examples{
1+1
}
