% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_state.R
\name{get_api_state}
\alias{get_api_state}
\title{Check API state}
\usage{
get_api_state(key = get_telraam_token())
}
\arguments{
\item{key}{the API key (set by the set_telraam_token function - default -, or directly filled).}
}
\value{
Boolean: TRUE if the API responds well, FALSE otherwise.
}
\description{
Return the state of the 'Telraam' API. Determine if updates can be made.
}
\examples{
my_token <- 'ThisIsNotAValidToken'
get_api_state(my_token)
}
