% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graph.R
\name{get_graph_subtitles}
\alias{get_graph_subtitles}
\title{Create subtitles for graphics.}
\usage{
get_graph_subtitles(
  segments = NULL,
  modes = NULL,
  directions = NULL,
  weekdays = NULL,
  hours = NULL
)
}
\arguments{
\item{segments}{Character vector. Selected road segment to text, no precision if NULL (default).}

\item{modes}{Character vector. Different modes of transportation selected (heavy, car, bike, pedestrian). Default: NULL}

\item{directions}{Character vector. Directions of the traffic (lft, rgt, both) choosen. Default to NULL.}

\item{weekdays}{Character vector. Weekdays choosen. Default to NULL.}

\item{hours}{Numeric vector. Hours choosen. Default to NULL.}
}
\value{
Character, with a description of all parameters filled, usable as subtitle for graphs.
}
\description{
Create subtitles for graphics.
}
\examples{
get_graph_subtitles(segments = c("Route1"),
  mode = "car", direction = "lft", weekdays = "monday", hours = 12:14)

}
\keyword{internal}
