% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_data.R
\name{impute_missing_data}
\alias{impute_missing_data}
\title{TODO}
\usage{
impute_missing_data(
  data,
  sensors_name = NULL,
  transport_type = "vehicle",
  threshold_uptime = 0.5
)
}
\arguments{
\item{data}{TBD}

\item{sensors_name}{TBD}

\item{transport_type}{TBD}

\item{threshold_uptime}{TBD}
}
\value{
TBD
}
\description{
TODO
}
\examples{
1+1
}
