% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{replace_inactivity_period}
\alias{replace_inactivity_period}
\title{Replace incomplete data with NA,}
\usage{
replace_inactivity_period(enriched_data, successive_day, uptime_choice)
}
\arguments{
\item{enriched_data}{enriched data.frame containing all the data for all your sensors}

\item{successive_day}{Integer. Number of day choosen. Default to 2}

\item{uptime_choice}{Real. Uptime choosen. Default to 0.5}
}
\value{
enriched_data
}
\description{
Replace incomplete data with NA,
}
\examples{
replace_inactivity_period(traffic)
replace_inactivity_period(traffic,
                          uptime_choice=0.3,
                          successive_day=1)
}
