% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{retrieve_missing_data}
\alias{retrieve_missing_data}
\title{Retrieve hours with no data and replace incomplete data with NA,}
\usage{
retrieve_missing_data(
  enriched_data,
  date_range = NULL,
  segments = NULL,
  uptime_choice = 0.5,
  successive_day = 2
)
}
\arguments{
\item{enriched_data}{enriched data.frame containing all the data for all your sensors}

\item{date_range}{Date vector. example: c('2021-01-01','2022-01-01'). Full period if NULL (default).}

\item{segments}{Character vector. Selected road segment, all if NULL (default).}

\item{uptime_choice}{Real. Uptime choosen. Default to 0.5}

\item{successive_day}{Integer. Number of day choosen. Default to 2}
}
\value{
enriched_data
}
\description{
Retrieve hours with no data and replace incomplete data with NA,
}
\examples{
retrieve_missing_data(traffic)
retrieve_missing_data(traffic,
                      date_range = c('2022-07-01','2022-09-01'),
                      segment = 'RteVitre-06',
                      uptime_choice=0.3,
                      successive_day=1)
}
