% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{retrieve_missing_hours}
\alias{retrieve_missing_hours}
\title{Retrieve hours with no data}
\usage{
retrieve_missing_hours(enriched_data, uptime_choice)
}
\arguments{
\item{enriched_data}{enriched data.frame containing all the data for all your sensors}

\item{uptime_choice}{Real. Uptime choosen. Default to 0.5}
}
\value{
enriched_data
}
\description{
Retrieve hours with no data
}
\examples{
retrieve_missing_hours(traffic)
retrieve_missing_hours(traffic,
                        uptime_choice=0.3)
}
