% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_data.R
\name{validate_and_preprocess_data}
\alias{validate_and_preprocess_data}
\title{Impute Missing Data for Traffic Analysis}
\usage{
validate_and_preprocess_data(data, transport_type, sensors_id, BASE_VARS)
}
\arguments{
\item{data}{A data frame containing the traffic data.}

\item{transport_type}{Character. Type of transport to impute. Options are "car", "vehicle", "heavy", or "all". Default is "vehicle".}

\item{sensors_id}{Character vector. Id of the sensors to include in the analysis. Default is NULL (all sensors).}

\item{BASE_VARS}{TBD}
}
\value{
A data frame with imputed values for the specified transport type(s) and a new column indicating whether the values were imputed or original.
}
\description{
This function imputes missing data for traffic analysis using Random Forest models.
It can handle different types of transport (car, vehicle, heavy) and can impute all types simultaneously.
}
\details{
The function requires specific columns to be present in the input data:
day_of_month, hour, weekday, month, year, vacation, week_number, segment_id, date, car, heavy, and uptime.

Vehicle is calculated as the sum of cars and heavy vehicles providing a good estimate of the total traffic.

When transport_type is "all", the function imputes values for car, vehicle, and heavy separately.

The function uses the uptime column to determine which data points need imputation. Values with uptime below the threshold_uptime are considered missing and are imputed.

Time interval could be hourly or quarterly. The function will automatically detect the time interval based on the data and add a minute variable if its quarterly.
}
\examples{
traffic_imputed <- impute_missing_data(traffic,
  sensors_name = "RteVitre-06",
  transport_type = "vehicle",
  threshold_uptime = 0.5)
table(traffic_imputed$imputed)
}
