% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_data.R
\name{validate_and_preprocess_data}
\alias{validate_and_preprocess_data}
\title{Validate and preprocess data for traffic analysis}
\usage{
validate_and_preprocess_data(
  data,
  transport_type,
  sensors_id,
  base_vars,
  add_vars = NULL,
  threshold_uptime = 0.5
)
}
\arguments{
\item{data}{A data frame containing the traffic data.}

\item{transport_type}{Character. Type of transport to impute.}

\item{sensors_id}{Character vector. Id of the sensors to include in the analysis.}

\item{base_vars}{Character vector. Base variables required for the analysis.}
}
\value{
A preprocessed data frame ready for imputation.
}
\description{
This function validates and preprocesses the input data for traffic analysis.
}
\examples{
data <- validate_and_preprocess_data(data = traffic,
                                     transport_type = "car",
                                     sensors_id = 9000001844,
                                     base_vars = c("day_of_month", "hour", "weekday",
                                                  "month", "year", "vacation",
                                                  "week_number", "segment_id", "date"))
}
\keyword{internal}
