% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_write_update_data_comp.R
\name{write_update_data_comp}
\alias{write_update_data_comp}
\title{Write or update the sensor data in the data folder}
\usage{
write_update_data_comp(
  id_sensor,
  date1,
  date2,
  vacations = NULL,
  public_holidays = NULL
)
}
\arguments{
\item{id_sensor}{Numeric. ID of the sensor}

\item{date1}{Date. Start date "aaaa-mm-jj"}

\item{date2}{Date. End date "aaaa-mm-jj"}

\item{vacations}{vacation periods, set by default on the french ones}

\item{public_holidays}{public holidays list, set by default on the french ones}
}
\description{
This function writes or updates the sensor data in the data folder. It retrieves the data for the specified sensor between \code{date1} and \code{date2} (inclusive) using the \code{retrieve_sensor} function, and then converts certain columns to character strings before writing the data to a CSV file in the data folder.
}
